updated routes to use model.methods
