+++ server/db/models/review.js
@@ -1,5 +1,7 @@
 var mongoose = require('mongoose');
 
+require("../../../server/db/models/user");
+var User = require('mongoose').model('User');
 
 function minLength(str) {
 	return str.length > 40;
@@ -12,5 +14,17 @@ var schema = new mongoose.Schema({
 	user: 	{ type: mongoose.Schema.Types.ObjectId, ref: 'User', required: false }
 });
 
+schema.methods.findUser = function findUser(){
+	// var userFindsPromiseArr = [];
+	// console.log('this review',this);
+	var review = this;
+	// review.forEach(function(review){
+		// userFindsPromiseArr.push
+		// (
+	return	User.findOne({"_id":review.user}).exec();
+		// );
+	// });
+	// return userFindsPromiseArr;
+};
 
 mongoose.model('Review', schema);
